requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
Pygments==2.10.0
pytz==2021.1
sqlparse==0.4.2
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
Django==3.2.7
django-cors-headers==3.8.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==3.2
pycparser==2.20
Pygments==2.10.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
Django==3.2.7
django-cors-headers==3.8.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==3.2
pycparser==2.20
Pygments==2.10.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
altgraph==0.17
appdirs==1.4.4
asgiref==3.4.1
attrs==20.3.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
certifi==2020.12.5
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
distlib==0.3.1
Django==3.2.6
docutils==0.17.1
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
idna==3.2
jellyfish==0.8.4
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
lib50==3.0.1
openpyxl==3.0.7
pefile==2021.9.3
pexpect==4.8.0
pipenv==2020.11.15
platformdirs==2.3.0
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pypiwin32==223
pytz==2021.1
pywin32==301
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.26.0
six==1.15.0
sqlparse==0.4.1
submit50==3.1.0
termcolor==1.1.0
toml==0.10.2
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.4
